# 1.8x Faster Partial Caching - Faster Cache Keys